requirements.txt
streamlit
pandas
numpy
time
matplotlib.pyplot
graphviz